backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
tzdata==2025.3
